pyproject.toml
[project]
name = "negative-lol"
version = "0.1.0"
description = ""
authors = [
    {name = "Ashwin Senthilvasan",email = "ashwin.senthilvasan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "requests (>=2.32.3,<3.0.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic[email] (>=2.11.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
]

[tool.poetry]
packages = [{include = "negative_lol", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
